# registry/openapi/registry-openapi.json
{
  "components": {
    "schemas": {
...
      "AgreementsRequest": {
        "properties": {
          "privacy": {
            "default": false,
            "title": "Privacy",
            "type": "boolean"
          },
          "publisher": {
            "default": false,
            "title": "Publisher",
            "type": "boolean"
          },
          "tos": {
            "default": false,
            "title": "Tos",
            "type": "boolean"
          }
        },
        "title": "AgreementsRequest",
        "type": "object"
      },
...
      "ApiTokenRotateRequest": {
        "properties": {
          "expiresInDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresindays"
          }
        },
        "title": "ApiTokenRotateRequest",
        "type": "object"
      },
      "ApiTokensResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiTokenInfo"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ApiTokensResponse",
        "type": "object"
      },
...
      "DeviceLoginRequest": {
        "properties": {
          "clientName": {
            "default": "cli",
            "title": "Clientname",
            "type": "string"
          },
          "mode": {
            "default": "poll",
            "title": "Mode",
            "type": "string"
          },
          "registryOrigin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registryorigin"
          },
          "requestedScopes": {
            "default": "cli",
            "title": "Requestedscopes",
            "type": "string"
          }
        },
        "title": "DeviceLoginRequest",
        "type": "object"
      },
      "DeviceLoginResponse": {
        "properties": {
          "deviceCode": {
            "title": "Devicecode",
            "type": "string"
          },
          "expiresIn": {
            "title": "Expiresin",
            "type": "integer"
          },
          "interval": {
            "title": "Interval",
            "type": "integer"
          },
          "userCode": {
            "title": "Usercode",
            "type": "string"
          },
          "verificationUri": {
            "title": "Verificationuri",
            "type": "string"
          }
        },
        "required": [
          "deviceCode",
          "userCode",
          "verificationUri",
          "expiresIn",
          "interval"
        ],
        "title": "DeviceLoginResponse",
        "type": "object"
      },
      "DevicePollRequest": {
        "properties": {
          "deviceCode": {
            "title": "Devicecode",
            "type": "string"
          }
        },
        "required": [
          "deviceCode"
        ],
        "title": "DevicePollRequest",
        "type": "object"
      },
      "DevicePollResponse": {
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "tokenType": {
            "default": "bearer",
            "title": "Tokentype",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "username"
        ],
        "title": "DevicePollResponse",
        "type": "object"
      },
      "GithubExchangeRequest": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "GithubExchangeRequest",
        "type": "object"
      },
      "GithubExchangeResponse": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "token",
          "username"
        ],
        "title": "GithubExchangeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "publicKeys": {
            "items": {
              "$ref": "#/components/schemas/SignerKeyInfo"
            },
            "title": "Publickeys",
            "type": "array"
          },
          "role": {
            "default": "publisher",
            "title": "Role",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "publicKeys"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "SuspendRequest": {
        "properties": {
          "suspended": {
            "title": "Suspended",
            "type": "boolean"
          }
        },
        "required": [
          "suspended"
        ],
        "title": "SuspendRequest",
        "type": "object"
      },
...
      "UploadKeyRequest": {
        "properties": {
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "publicKey": {
            "title": "Publickey",
            "type": "string"
          }
        },
        "required": [
          "publicKey"
        ],
        "title": "UploadKeyRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "OpenAgentHub Registry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/users/{user_id}/suspend": {
      "post": {
        "operationId": "admin_suspend_user_api_v1_admin_users__user_id__suspend_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Suspend User"
      }
    },
...
    "/api/v1/auth/approve": {
      "post": {
        "operationId": "approve_device_api_v1_auth_approve_post",
        "parameters": [
          {
            "in": "query",
            "name": "user_code",
            "required": true,
            "schema": {
              "title": "User Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Device"
      }
    },
    "/api/v1/auth/devices": {
      "post": {
        "operationId": "create_device_api_v1_auth_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Device"
      }
    },
    "/api/v1/auth/devices/token": {
      "post": {
        "operationId": "poll_device_api_v1_auth_devices_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll Device"
      }
    },
    "/api/v1/auth/github": {
      "post": {
        "operationId": "github_login_api_v1_auth_github_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GithubExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubExchangeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Github Login"
      }
    },
...
    "/api/v1/keys": {
      "post": {
        "operationId": "upload_key_api_v1_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Key"
      }
    },
    "/api/v1/keys/{key_id}": {
      "delete": {
        "operationId": "revoke_key_api_v1_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Key"
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me"
      }
    },
...
    "/api/v1/me/agreements": {
      "get": {
        "operationId": "my_agreements_api_v1_me_agreements_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "My Agreements"
      },
      "post": {
        "operationId": "update_agreements_api_v1_me_agreements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgreementsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agreements"
      }
    },
...
    "/api/v1/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_my_session_api_v1_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke My Session"
      }
    },
    "/api/v1/tokens": {
      "get": {
        "operationId": "my_tokens_api_v1_tokens_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokensResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Tokens"
      },
      "post": {
        "operationId": "create_token_api_v1_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Token"
      }
    },
    "/api/v1/tokens/{token_id}": {
      "delete": {
        "operationId": "revoke_token_api_v1_tokens__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "title": "Token Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Token"
      }
    },
    "/api/v1/tokens/{token_id}/rotate": {
      "post": {
        "operationId": "rotate_token_api_v1_tokens__token_id__rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "title": "Token Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenRotateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rotate Token"
      }
    },
    "/auth/github/callback": {
      "get": {
        "operationId": "github_callback_auth_github_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Github Callback"
      }
    },
    "/auth/github/start": {
      "get": {
        "operationId": "github_start_auth_github_start_get",
        "parameters": [
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "schema": {
              "title": "Redirect Uri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Github Start"
      }
    },
...
  }
}